Add task base classes
This PR grew as I was writing it, and now tries out a number of new
things:
1. Exposed base classes. Sets us on a path for better documentation,
   a more "introspectable" library, and allow sub-classing.
2. Enable `from_preset()` on base classes for any subclass preset. This
   gives us similar functionality to "auto classes" in huggingface,
   without the extra overhead of needing a new symbol.
3. An ability to register new tasks/backbones/tokenizers from out of
   tree code with `keras_nlp.utils.register_preset_class()`.

`1.` seems like something we probably need.

`2.` adds a bit of complexity, but seems super practical by allowing
portable code for, say, instantiating a tokenizer:
```
tokenizer = keras_nlp.models.Tokenizer.from_preset(
    "kaggle://someuser/somemodel/keras/somefinetune",
)
```

`3.` seems like something we would want eventually as a consequence of
`2.`, but would definitely put us on the hook more for documenting and
sticking by a way to write model subclasses. We could for sure punt on
this.

Try a colab:
https://colab.research.google.com/gist/mattdangerw/da885f050fa8baef9b4f9a4ec68d6567/kerasnlp-base-classes.ipynb